Enhanced TypeUtil
